backend/requirements.txt
langchain-community==0.3.16
langchain-openai==0.3.3
langchain-core==0.3.33
langchain-postgres==0.0.12
chromadb==0.6.3
langchain-chroma==0.2.1
fastapi==0.115.8
langserve==0.3.1
sse_starlette==2.2.1
pymupdf==1.25.3
psycopg-binary==3.2.9
pydantic==2.10.6
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6